Add files via upload
In 1973, the world was introduced to the simplistic, but rather controversial Black and Scholes model. I have created a function, such that call and put options can be priced. Additionally, with the assistance of the 'Black_Scholes_Plots function', it is possible to visualise how option prices are impacted by certain factors (volatility, strike prices, etc).